requirements.txt
aniso8601==8.1.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
flasgger==0.9.5
Flask==1.1.2
Flask-RESTful==0.3.8
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
logging==0.4.9.6
MarkupSafe==1.1.1
mistune==0.8.4
nltk==3.5
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
tqdm==4.56.2
urllib3==1.26.3
Werkzeug==1.0.1
